move market maker validator to separate package
