+++ d3a_interface/validators/market_maker_validator.py
@@ -1,0 +1,30 @@
+"""
+Copyright 2018 Grid Singularity
+This file is part of D3A.
+
+This program is free software: you can redistribute it and/or modify it under the terms of the
+GNU General Public License as published by the Free Software Foundation, either version 3 of the
+License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along with this program. If
+not, see <http://www.gnu.org/licenses/>.
+"""
+from d3a_interface.exceptions import D3ADeviceException
+from d3a_interface.validators import utils
+
+
+def validate_market_maker(**kwargs):
+    utils.validate_energy_rate(**kwargs)
+    if "energy_rate_profile" in kwargs and kwargs["energy_rate_profile"] is not None and \
+            ("energy_rate_profile_uuid" not in kwargs or
+             kwargs["energy_rate_profile_uuid"] is None):
+        raise D3ADeviceException(
+            {"misconfiguration": [f"energy_rate_profile must have a uuid."]})
+    if "grid_connected" in kwargs and kwargs["grid_connected"] is not None and \
+            not isinstance(kwargs["grid_connected"], bool):
+        raise D3ADeviceException(
+            {"misconfiguration": [f"grid_connected must be a boolean value."]})
